requirements.txt
mysql-connector-python==8.0.28
protobuf==3.20.1
